+++ lib/server/server.dart
@@ -9,15 +9,21 @@ import 'package:connectars/service/log.dart';
 import 'package:dotenv/dotenv.dart';
 
 void run() async {
-  var server = await HttpServer.bind(env['SOCKET_HOST'], int.parse(env['SOCKET_PORT']));
+  var server =
+      await HttpServer.bind(env['SOCKET_HOST'], int.parse(env['SOCKET_PORT']));
 
   await for (HttpRequest request in server) {
-    LogService().log('Time: ' + DateTime.now().toUtc().toString(), type: LogService.typeRequest);
-    LogService().log('Route: ' + request.uri.path, type: LogService.typeRequest);
+    LogService().log('Time: ' + DateTime.now().toUtc().toString(),
+        type: LogService.typeRequest);
+    LogService()
+        .log('Route: ' + request.uri.path, type: LogService.typeRequest);
     LogService().log('Method: ' + request.method, type: LogService.typeRequest);
-    LogService().log('Headers: ' + request.headers.toString(), type: LogService.typeRequest);
-    LogService().log('Query: ' + request.uri.queryParameters.toString(), type: LogService.typeRequest);
-    LogService().log('Body: ' + await utf8.decodeStream(request), type: LogService.typeRequest);
+    LogService().log('Headers: ' + request.headers.toString(),
+        type: LogService.typeRequest);
+    LogService().log('Query: ' + request.uri.queryParameters.toString(),
+        type: LogService.typeRequest);
+    LogService().log('Body: ' + await utf8.decodeStream(request),
+        type: LogService.typeRequest);
 
     final response = request.response;